requirements-reporter.txt
# 金融报告系统额外依赖
# Additional dependencies for Financial Reporter System

# Web 服务器
flask>=3.0.0

# Markdown 渲染
markdown>=3.5.0

# 定时任务
schedule>=1.2.0

# 所有 Mini Agent 的基础依赖
pydantic>=2.0.0
pyyaml>=6.0.0
httpx>=0.27.0
mcp>=1.0.0
requests>=2.31.0
tiktoken>=0.5.0
prompt-toolkit>=3.0.0
anthropic>=0.39.0
openai>=1.57.4
